requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2==2.9.5
requests==2.28.2
